Restructure files in src
