+++ src/discord/discordWatcher/onClientReady.ts
