+++ src/discord/interactions/core/registry.ts
