+++ src/discord/interactions/typeGuards.ts
