+++ src/discord/messageEvaluator/util.ts
